+++ apm/lib/collections.js
@@ -8,7 +8,6 @@ MapReduceProfileConfig = new Mongo.Collection('mapReduceProfileConfig');
 ErrorTraces = new Mongo.Collection('errorTraces');
 ErrorMetrics = new Mongo.Collection('errorMetrics');
 AppStats = new Mongo.Collection('appStats');
-ProdStats = new Mongo.Collection('prodStats');
 MethodsMetrics = new Mongo.Collection('methodsMetrics');
 MethodTraces = new Mongo.Collection('methodsTraces');
 RawErrorMetrics = new Mongo.Collection('rawErrorMetrics');
@@ -21,65 +20,24 @@ SystemMetrics = new Mongo.Collection('systemMetrics');
 RmaLogs = new Mongo.Collection('rmaLogs');
 
 if (Meteor.isServer) {
+  const metricsIndex = {
+    'value.res': 1,
+    'value.appId': 1,
+    'value.startTime': -1
+  };
+  SystemMetrics.rawCollection().createIndex(metricsIndex);
+  MethodsMetrics.rawCollection().createIndex(metricsIndex);
+  PubMetrics.rawCollection().createIndex(metricsIndex);
+  ErrorMetrics.rawCollection().createIndex(metricsIndex);
+
   AppStats.rawCollection().createIndex({
     'value.res': 1,
     'value.appId': 1,
     'value.host': 1,
-    'value.startTime': 1
+    'value.startTime': -1
   });
-  SystemMetrics.rawCollection().createIndex({
-    'value.res': 1,
-    'value.appId': 1,
-    'value.host': 1,
-    'value.startTime': 1
-  });
+
   MethodTraces.rawCollection().createIndex({ appId: 1, host: 1, startTime: 1 });
   PubTraces.rawCollection().createIndex({ appId: 1, host: 1, startTime: 1 }, { background: true });
   ErrorTraces.rawCollection().createIndex({ appId: 1, host: 1, startTime: 1 });
-  MethodsMetrics.rawCollection().createIndex({
-    'value.res': 1,
-    'value.appId': 1,
-    'value.host': 1,
-    'value.startTime': 1
-  });
-  PubMetrics.rawCollection().createIndex({
-    'value.res': 1,
-    'value.appId': 1,
-    'value.host': 1,
-    'value.startTime': 1
-  });
-  ErrorMetrics.rawCollection().createIndex({
-    'value.res': 1,
-    'value.appId': 1,
-    'value.host': 1,
-    'value.startTime': 1
-  });
-  ProdStats.rawCollection().createIndex({
-    appId: 1,
-    metric: 1
-  });
-  RawErrorMetrics.rawCollection().createIndex({
-    'value.res': 1,
-    'value.appId': 1,
-    'value.host': 1,
-    'value.startTime': 1
-  });
-  RawMethodsMetrics.rawCollection().createIndex({
-    'value.res': 1,
-    'value.appId': 1,
-    'value.host': 1,
-    'value.startTime': 1
-  });
-  RawPubMetrics.rawCollection().createIndex({
-    'value.res': 1,
-    'value.appId': 1,
-    'value.host': 1,
-    'value.startTime': 1
-  });
-  RawSystemMetrics.rawCollection().createIndex({
-    'value.res': 1,
-    'value.appId': 1,
-    'value.host': 1,
-    'value.startTime': 1
-  });
 }