+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+## [1.0.2] - 2017-11-10
+### Changed
+- Optimized indexes of metrics collections.
+- package.json name and description.
+### Removed
+- Unused collection ProdStats.
 
 ## [1.0.1] - 2017-11-09
 ### Added